Refactor the Forecaster logic into its own Class.
