+++ src/include/self_driving/planning/pilot.h
@@ -15,6 +15,7 @@
 #include "common/managed_pointer.h"
 #include "execution/exec_defs.h"
 #include "metrics/query_trace_metric.h"
+#include "self_driving/forecasting/forecaster.h"
 #include "self_driving/forecasting/workload_forecast.h"
 #include "self_driving/planning/action/action_defs.h"
 
@@ -82,27 +83,6 @@ class Pilot {
    */
   static constexpr bool WHAT_IF = true;
 
-  /** Describes how the workload forecast should be initialized */
-  enum class WorkloadForecastInitMode : uint8_t {
-    /**
-     * Construct the workload forecast solely from data stored in internal tables.
-     * Passes data read from internal tables to perform inference
-     */
-    INTERNAL_TABLES_WITH_INFERENCE,
-
-    /**
-     * Construct the workload forecast by inferencing data located on disk.
-     * The inference result is not stored to internal tables in this mode.
-     */
-    DISK_WITH_INFERENCE,
-
-    /**
-     * Construct the workload forecast directly from data on disk.
-     * No inference is performed in this case.
-     */
-    DISK_ONLY
-  };
-
   /**
    * Constructor for Pilot
    * @param ou_model_save_path OU model save path
@@ -116,7 +96,7 @@ class Pilot {
    * @param txn_manager transaction manager
    * @param query_exec_util query execution utility for the pilot to use
    * @param task_manager task manager to submit internal jobs to
-   * @param workload_forecast_interval Interval used in the forecastor
+   * @param workload_forecast_interval Interval used in the Forecaster
    * @param sequence_length Length of a planning sequence
    * @param horizon_length Length of the planning horizon
    */
@@ -142,58 +122,20 @@ class Pilot {
   common::ManagedPointer<modelserver::ModelServerManager> GetModelServerManager() { return model_server_manager_; }
 
   /**
-   * Rerieve segment information
-   * TODO(wz2): Addressing clustering based on this data will be left for the future.
-   *
-   * @param bounds (inclusive) bounds of the time range
-   * @param success [out] indicator of whether query succeeded or not
-   * @return segment information
-   */
-  std::unordered_map<int64_t, std::vector<double>> GetSegmentInformation(std::pair<uint64_t, uint64_t> bounds,
-                                                                         bool *success);
-
-  /**
-   * Retrieve workload metadata
-   * @param bounds (inclusive) bounds of the time range to pull data
-   * @param out_metadata Query Metadata from metrics
-   * @param out_params Query parameters fro mmetrics
-   * @return pair where first is metadata and second is flag of success
-   */
-  std::pair<selfdriving::WorkloadMetadata, bool> RetrieveWorkloadMetadata(
-      std::pair<uint64_t, uint64_t> bounds,
-      const std::unordered_map<execution::query_id_t, metrics::QueryTraceMetadata::QueryMetadata> &out_metadata,
-      const std::unordered_map<execution::query_id_t, std::vector<std::string>> &out_params);
-
-  /**
-   * Record the workload forecast to the internal tables
-   * @param timestamp Timestamp to record forecast at
-   * @param prediction Forecast model prediction
-   * @param metadata Metadata about the queries
-   */
-  void RecordWorkloadForecastPrediction(uint64_t timestamp, const selfdriving::WorkloadForecastPrediction &prediction,
-                                        const WorkloadMetadata &metadata);
-
-  /**
-   * Loads workload forecast information
-   * @param mode Mode to initialize forecast information
-   */
-  void LoadWorkloadForecast(WorkloadForecastInitMode mode);
-
-  /**
    * Performs Pilot Logic, load and execute the predicted queries while extracting pipeline features
    */
   void PerformPlanning();
-
-  /**
-   * Performs training of the forecasting model
-   */
-  void PerformForecasterTrain();
 
   /**
    * Search for and apply the best action for the current timestamp
    * @param best_action_seq pointer to the vector to be filled with the sequence of best actions to take at current time
    */
   void ActionSearch(std::vector<std::pair<const std::string, catalog::db_oid_t>> *best_action_seq);
+
+  /**
+   * Performs training of the forecasting model
+   */
+  void PerformForecasterTrain() { forecaster_.PerformTraining(); };
 
  private:
   /**
@@ -220,17 +162,8 @@ class Pilot {
                        std::map<uint32_t, uint64_t> *segment_to_offset,
                        std::vector<std::vector<double>> *interference_result_matrix);
 
-  /**
-   * Computes the valid range of data to be pulling from the internal tables.
-   * @param now Current timestamp of the planning/training
-   * @param train Whether data is for training or inference
-   * @return inclusive start and end bounds of data to query
-   */
-  std::pair<uint64_t, uint64_t> ComputeTimestampDataRange(uint64_t now, bool train);
-
   std::string ou_model_save_path_;
   std::string interference_model_save_path_;
-  std::string forecast_model_save_path_;
   common::ManagedPointer<catalog::Catalog> catalog_;
   common::ManagedPointer<metrics::MetricsThread> metrics_thread_;
   common::ManagedPointer<modelserver::ModelServerManager> model_server_manager_;
@@ -239,9 +172,7 @@ class Pilot {
   common::ManagedPointer<transaction::TransactionManager> txn_manager_;
   std::unique_ptr<util::QueryExecUtil> query_exec_util_;
   common::ManagedPointer<task::TaskManager> task_manager_;
-  uint64_t workload_forecast_interval_{1000000};
-  uint64_t sequence_length_{10};
-  uint64_t horizon_length_{30};
+  Forecaster forecaster_;
   uint64_t action_planning_horizon_{5};
   uint64_t simulation_number_{20};
   friend class noisepage::selfdriving::PilotUtil;
